// src/app/model/pokemondetails.model.ts
export class PokemonDetails {

   name: string;
   type: string;
   height: string;
   weight: string;
   signature: string;
   experience: string;

   constructor() {
     /*this.name = data.name;
     this.type = data.type;
     this.height = data.height;
     this.weight = data.weight;
     this.signature = data.signature;
     this.experience = data.experience;*/
   }
}